Fix bug 309863
The crash occurs when AI is playing or Auto attack is going on and user wants to see country details from context menu. Both country and fight details are shown in the right dialog interfering with each other. Some safety code is added to prevent new crashes.
The complete solution is to disable detailed country view from the context menu when AI is playing or Auto Attack is under way. This separates appearing of two views at the same time in the right dialog.

BUGS: 309863
REVIEW: 108068
FIXED-IN: 4.10.1